moved hook into hooks file
