--- src/pages/proposal/ProposalDisplay/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useVoteConfirmation'
